Feature: Started programming on add movie
This features implies that the user gives list containing the,
where he/she saw it and an evaluation (features to be extended).

The function then checks if there is already a csv and if yes
it appends this information to it. And if not, it makes a new
csv with this information.